Extract infrastructure configuration extension methods in Shared Kernel to a separate file
